Added S3 storage implementation based on async aws

+++ src/bridge/filesystem/async-aws/src/Flow/Filesystem/Bridge/AsyncAWS/AsyncAWSS3Filesystem.php
@@ -8,17 +8,33 @@ use AsyncAws\S3\Exception\NoSuchKeyException;
 use AsyncAws\S3\S3Client;
 use Flow\Filesystem\Path\Filter;
 use Flow\Filesystem\Path\Filter\KeepAll;
-use Flow\Filesystem\{DestinationStream, FileStatus, Filesystem, Path, Protocol, SourceStream};
+use Flow\Filesystem\{DestinationStream,
+    Exception\InvalidArgumentException,
+    Exception\RuntimeException,
+    FileStatus,
+    Filesystem,
+    Path,
+    Protocol,
+    SourceStream};
 
 final class AsyncAWSS3Filesystem implements Filesystem
 {
-    public function __construct(private readonly S3Client $s3Client, private readonly Options $options)
+    public function __construct(private readonly string $bucket, private readonly S3Client $s3Client, private readonly Options $options)
     {
-
+        if ($bucket === '') {
+            throw new InvalidArgumentException('Bucket name can not be empty');
+        }
     }
 
     public function appendTo(Path $path) : DestinationStream
     {
+        if ($path->isEqual($this->getSystemTmpDir())) {
+            throw new RuntimeException('Cannot write to system tmp directory');
+        }
+
+        $this->protocol()->validateScheme($path);
+
+        return AsyncAWSS3DestinationStream::openAppend($this->s3Client, $this->bucket, $path, $this->options->blockFactory(), $this->options->partSize());
     }
 
     public function getSystemTmpDir() : Path
@@ -28,12 +44,62 @@ final class AsyncAWSS3Filesystem implements Filesystem
 
     public function list(Path $path, Filter $pathFilter = new KeepAll()) : \Generator
     {
-        // TODO: Implement list() method.
+        if ($path->rootDirectoryName() === null) {
+            throw new InvalidArgumentException('Root directory name can not be empty for S3 destination stream since it represents bucket name. Example: aws-s3://bucket_name/file.txt');
+        }
+
+        $this->protocol()->validateScheme($path);
+
+        if ($path->isPattern()) {
+            $prefix = \ltrim($path->staticPart()->path(), DIRECTORY_SEPARATOR);
+        } else {
+            $prefix = \ltrim($path->path(), DIRECTORY_SEPARATOR);
+        }
+
+        $continuationToken = null;
+
+        do {
+            $result = $this->s3Client->listObjectsV2([
+                'Bucket' => $this->bucket,
+                'Prefix' => $prefix,
+                'ContinuationToken' => $continuationToken,
+            ]);
+
+            foreach ($result->getContents() as $object) {
+                /** @psalm-suppress PossiblyNullArgument */
+                $objectPath = new Path($path->protocol()->scheme() . DIRECTORY_SEPARATOR . \ltrim($object->getKey(), DIRECTORY_SEPARATOR), $path->options());
+                $objectFileStatus = new FileStatus($objectPath, (bool) $objectPath->extension());
+
+                if ($path->isPattern() && !$path->matches($objectPath)) {
+                    continue;
+                }
+
+                if ($pathFilter->accept($objectFileStatus)) {
+                    yield $objectFileStatus;
+                }
+            }
+
+            $continuationToken = $result->getNextContinuationToken();
+        } while ($continuationToken);
     }
 
     public function mv(Path $from, Path $to) : bool
     {
-        // TODO: Implement mv() method.
+        $this->protocol()->validateScheme($from);
+        $this->protocol()->validateScheme($to);
+
+        $this->s3Client->copyObject([
+            'Bucket' => $this->bucket,
+            'Key' => ltrim($to->path(), '/'),
+            'CopySource' => $this->bucket . '/' . ltrim($from->path(), '/'),
+        ]);
+
+        $this->s3Client->deleteObject([
+            'Bucket' => $this->bucket,
+            'Key' => ltrim($from->path(), '/'),
+        ]);
+
+        return true;
     }
 
     public function protocol() : Protocol
@@ -43,7 +109,7 @@ final class AsyncAWSS3Filesystem implements Filesystem
 
     public function readFrom(Path $path) : SourceStream
     {
-        return new S3SourceStream($path, $this->s3Client);
+        return new AsyncAWSS3SourceStream($path, $this->bucket, $this->s3Client);
     }
 
     public function rm(Path $path) : bool
@@ -52,16 +118,55 @@ final class AsyncAWSS3Filesystem implements Filesystem
             return false;
         }
 
-        try {
-            $this->s3Client->deleteObject([
-                'Bucket' => $path->rootDirectoryName(),
-                'Key' => $path->skipDirectories(1)?->path(),
-            ]);
+        if ($path->isPattern()) {
+            $deletedCount = 0;
 
-            return true;
-        } catch (NoSuchKeyException $e) {
-            return false;
+            foreach ($this->list($path) as $fileStatus) {
+                $this->s3Client->deleteObject([
+                    'Bucket' => $this->bucket,
+                    'Key' => ltrim($fileStatus->path->path(), '/'),
+                ]);
+
+                $deletedCount++;
+            }
+
+            return (bool) $deletedCount;
         }
+
+        try {
+            $headObject = $this->s3Client->headObject([
+                'Bucket' => $this->bucket,
+                'Key' => ltrim($path->path(), '/'),
+            ]);
+            $headObject->resolve();
+        } catch (NoSuchKeyException $e) {
+            /**
+             * Since AzureS3 doesn't have a concept of folders, before we check if the intention is not to delete
+             * entire path, like for example aws-s3://nested/folder we need to first add / at the end, to accidentally
+             * not delete files that would also match the prefix, like: aws-s3://nested/folder_but_file.txt.
+             */
+            $folderPath = new Path(\trim($path->uri(), DIRECTORY_SEPARATOR) . DIRECTORY_SEPARATOR, $path->options());
+
+            $deletedCount = 0;
+
+            foreach ($this->list($folderPath) as $fileStatus) {
+                $this->s3Client->deleteObject([
+                    'Bucket' => $this->bucket,
+                    'Key' => ltrim($fileStatus->path->path(), '/'),
+                ]);
+                $deletedCount++;
+            }
+
+            return (bool) $deletedCount;
+        }
+
+        $this->s3Client->deleteObject([
+            'Bucket' => $this->bucket,
+            'Key' => ltrim($path->path(), '/'),
+        ]);
+
+        return true;
+
     }
 
     public function status(Path $path) : ?FileStatus
@@ -79,8 +184,8 @@ final class AsyncAWSS3Filesystem implements Filesystem
 
             try {
                 $headObject = $this->s3Client->headObject([
-                    'Bucket' => $path->rootDirectoryName(),
-                    'Key' => $path->skipDirectories(1)?->path(),
+                    'Bucket' => $this->bucket,
+                    'Key' => ltrim($path->path(), '/'),
                 ]);
                 $headObject->resolve();
 
@@ -110,6 +215,12 @@ final class AsyncAWSS3Filesystem implements Filesystem
 
     public function writeTo(Path $path) : DestinationStream
     {
-        return S3DestinationStream::openBlank($this->s3Client, $path, $this->options->blockFactory(), $this->options->partSize());
+        if ($path->isEqual($this->getSystemTmpDir())) {
+            throw new RuntimeException('Cannot write to system tmp directory');
+        }
+
+        $this->protocol()->validateScheme($path);
+
+        return AsyncAWSS3DestinationStream::openBlank($this->s3Client, $this->bucket, $path, $this->options->blockFactory(), $this->options->partSize());
     }
 }
